+++ public/js/decks.js
@@ -186,30 +186,35 @@ export const renderDeck = async (id) => {
   username.classList.add('username');
   username.textContent = `User : ${deck.userName}, Created at ${new Date(deck.createdAt).toString()}`;
 
-  const deleteBtn = document.createElement('button');
-  deleteBtn.classList.add('delete-btn', 'btn', 'btn-danger', 'btn-sm', 'float-end', 'modal-btn');
-  deleteBtn.textContent = 'Delete';
-
-  deleteBtn.addEventListener('click', async () => {
-    console.log(localStorage.getItem('user'));
-    const confirmation = await showModal('Are you sure you want to delete this deck?', 'This action cannot be undone.');
-    if (confirmation) {
-      showAlert('success', 'Deck deleted successfully');
-      await deleteDeck(id);
-      location.assign('/decks');
-    } else {
-      showAlert('info', 'Deletion canceled');
-      console.log('Deletion canceled');
-    }
-  });
-
   userInfo.appendChild(username);
   cardCover.appendChild(coverImg);
   card.appendChild(cardCover);
   card.appendChild(nameHeader);
   card.appendChild(archetype);
   card.appendChild(userInfo);
-  card.appendChild(deleteBtn);
+
+  const curUser = JSON.parse(localStorage.getItem('user'));
+  console.log(curUser);
+  if (curUser.id === deck.userId) {
+    const deleteBtn = document.createElement('button');
+    deleteBtn.classList.add('delete-btn', 'btn', 'btn-danger', 'btn-sm', 'float-end', 'modal-btn');
+    deleteBtn.textContent = 'Delete';
+
+    deleteBtn.addEventListener('click', async () => {
+      console.log(localStorage.getItem('user'));
+      const confirmation = await showModal('Are you sure you want to delete this deck?', 'This action cannot be undone.');
+      if (confirmation) {
+        showAlert('success', 'Deck deleted successfully');
+        await deleteDeck(id);
+        location.assign('/decks');
+      } else {
+        showAlert('info', 'Deletion canceled');
+        console.log('Deletion canceled');
+      }
+    });
+
+    card.appendChild(deleteBtn);
+  }
 
   deckCardContainer.appendChild(card);
